[Refactor] 에러처리를 위한 타입들 코어 패키지로 리팩토링
