+++ app/src/main/java/com/hara/kaera/core/ApiResult.kt
@@ -1,0 +1,6 @@
+package com.hara.kaera.core
+
+sealed class ApiResult<out T> {
+    data class Success<out T>(val data: T) : ApiResult<T>()
+    data class Error<T>(val error: ErrorType) : ApiResult<T>()
+}
