+++ app/src/main/java/com/hara/kaera/core/ErrorCode.kt
@@ -1,4 +1,4 @@
-package com.hara.kaera.data.util
+package com.hara.kaera.core
 
 object ErrorCodes {
